Updated Rooms.cpp to recognize item encounters. Updated descriptions in Items.cpp for weapons. Updated Items.cpp to have armor choices.

+++ Enemies.cpp
@@ -7,7 +7,7 @@
 
 using namespace std;
 
-Enemy::Enemy(string name, float health, float baseDamage)
+Enemy::Enemy(string name, int health, int baseDamage)
 	: Player(name, health, baseDamage)
 { 
 }
